Use main in github workflow (#1380)
Default branch is now main, so all the PRs are based on main, which
means the actions aren't currently triggered, since they only run on
push to master. Update it to main.

Since we have the main->master mirror, we should still be testing, but
it is a bit delayed (only after the mirror I think, and not on the PR).